tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/043.html

<!doctype html>
<title>document.write entity</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
t.step(function() {
  document.write("<body><span>&");
});
</script>notabc</span>
<script>
t.step(function() {
  assert_equals(document.body.childNodes[0].textContent, "\u00ACabc");
})
t.done();
</script>
<div id="log"></div>
